User commit

+++ src/main/java/com/example/demo/controller/ValidationController.java
@@ -23,7 +23,7 @@ public class ValidationController{
    }
    
    @GetMapping("/valfind/{id}")
-   public ValidationEntity find(@PathVariable int id){
+   public ValidationEntity find(@PathVariable Long id){
     return ser.getdata(id);
    }
